refactor(DFS/BFS): splitted logics to each files
